Move to vpn-impl module
